backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.24.3
opencv-python==4.8.1.78
ultralytics==8.0.186
pillow==10.1.0
geopy==2.3.0
boto3==1.28.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
aiofiles==23.2.1
openpyxl==3.1.2
exifread==3.0.0
torch==2.0.1
yandexcloud==0.37.0